+++ src/DataSet_GridFlt.h
@@ -19,19 +19,19 @@ class DataSet_GridFlt : public DataSet_3D {
     void Info()                          const { return;                     }
     void WriteBuffer(CpptrajFile&,SizeArray const&) const;
     // ----- DataSet_3D functions ----------------
-    int Allocate3D(size_t x,size_t y,size_t z) { return grid_.resize(x,y,z); }
-    double GetElement(int x,int y,int z) const { return (double)grid_.element(x,y,z); }
-    double operator[](size_t idx)        const { return (double)grid_[idx];  }
+    int Allocate3D(size_t x, size_t y, size_t z)          { return grid_.resize(x,y,z); }
+    double GetElement(size_t x, size_t y, size_t z) const { return (double)grid_.element(x,y,z); }
+    double operator[](size_t idx)                   const { return (double)grid_[idx];  }
     size_t NX() const { return grid_.NX(); }
     size_t NY() const { return grid_.NY(); }
     size_t NZ() const { return grid_.NZ(); }
     /// \return grid index
-    long int CalcIndex(int i, int j, int k) const { return grid_.CalcIndex(i,j,k); }
-    void ReverseIndex(long int n, int& i, int& j, int& k) const
+    long int CalcIndex(size_t i, size_t j, size_t k) const { return grid_.CalcIndex(i,j,k); }
+    void ReverseIndex(long int n, size_t& i, size_t& j, size_t& k) const
       { return grid_.ReverseIndex(n,i,j,k); }
     void UpdateVoxel(long int i, double val) { grid_[i] += (float)val; }
     // -------------------------------------------
-    void SetElement(int x,int y,int z,float v) { grid_.setGrid(x,y,z,v);     }
+    void SetElement(size_t x,size_t y,size_t z,float v) { grid_.setGrid(x,y,z,v);     }
     /// Type definition of iterator over grid elements.
     typedef Grid<float>::iterator iterator;
     iterator begin() { return grid_.begin(); }
@@ -40,29 +40,29 @@ class DataSet_GridFlt : public DataSet_3D {
     inline long int Increment(Vec3 const&, float);
     inline long int Increment(const double*, float);
     /// Increment grid bin by given value.
-    inline long int Increment(int,int,int,float);
+    inline long int Increment(size_t,size_t,size_t,float);
     /// \return grid value at specified bin.
-    float GridVal(int x,int y,int z)        const { return grid_.element(x,y,z);   }
+    float GridVal(size_t x,size_t y,size_t z)        const { return grid_.element(x,y,z);   }
   private:
     Grid<float> grid_;
 };
 // ----- INLINE FUNCTIONS ------------------------------------------------------
 // DataSet_GridFlt::Increment()
 long int DataSet_GridFlt::Increment(Vec3 const& xyz, float f) {
-  int i,j,k;
+  size_t i,j,k;
   if (CalcBins(xyz[0],xyz[1],xyz[2],i,j,k))
     return grid_.incrementBy(i,j,k,f);
   return -1L; 
 }
 // DataSet_GridFlt::Increment()
 long int DataSet_GridFlt::Increment(const double* xyz, float f) {
-  int i,j,k;
+  size_t i,j,k;
   if (CalcBins(xyz[0],xyz[1],xyz[2],i,j,k))
     return grid_.incrementBy(i,j,k,f);
   return -1L;
 }
 // DataSet_GridFlt::Increment()
-long int DataSet_GridFlt::Increment(int i, int j, int k, float f) {
+long int DataSet_GridFlt::Increment(size_t i, size_t j, size_t k, float f) {
   return grid_.incrementBy(i,j,k,f);
 }
 #endif
